Tweak linter config and fix all lint errors (#9)
* Add @flimbul/wotan and @fimbul/valtry to devDependencies

* Remove "check-else" from "one-line" options

Rationale: It was eating important documenting comments placed before each branch of a long series of if/elseif. I'll consider a style for these comments other than '}' + newline + // comment + newline + 'else if(...)' but for now I am removing the rule in the interest of getting this done quicker

* WIP fix lint errors

Mostly automatic with --fix but I'm also going through manually to fix some of the less well handled cases

* Removed "member-ordering" rule

To hopefully one day be replaced with a less broken rule.

The rationale: The linter doesn't know what the heck to do with getters and setters and screws up the member order in classes that have them. Apparently related to https://www.github.com/palantir/tslint/issues/3965

* Fix more lint errors

* More fixing of member order related lint errors

....though strictly speaking these aren't really lint errors anymore since I removed the rule for misbehaving. But still these things should be fixed

* Add "allow-empty-functions" to "no-empty" linter option

The rationale: () => {} is the best way to write a nop anonymous function and previously the linter would throw up an error on these

* Set "ordered-imports" option "grouped-imports" to false in lint config

Rationale: It was filling my log with errors I don't understand even after reading the docs for this rule, --fix doesn't fix them, and I don't have enough patience to puzzle it out right now.

* Fix the last of the remaining lint errors