Attempt to optimize TextureLibrary
Refactor TextureLibrary so that it does not get all of the textures at 
once, but instead, only gets textures when they are needed.
Remove getTextureLibrary function and textureLibrary global from 
lumpTypeView.ts
Re-write WADTexture.isTransparent so that it iterates through the 
patches, builds an alpha map, and checks whether any pixel on that map 
is less than 255.