Make further general optimizations
Allow textures in the texture library to be set to null. This lets the 
texture library cache "-" as null, so that it doesn't search the texture 
lists or flats for it.
Remove dead code in TextureLibrary.
Cache RGBA pixel data in TextureLibrary for easy access.
Add DataManager class to lumpTypeView.ts to allow re-use of WAD file 
lists and the texture library to be re-used between views.