Initial implementation of sector polygon joining, and bug fixes
Fix some edges being added twice to the sector polygon.
Add badVertex argument to findNextEdgeFor, and vertexRefCount argument to findNextGoodEdge.
If a "bad" vertex is given to findNextEdgeFor, the method tries to avoid finding an edge that has that particular vertex attached.
Copy this.vertexRefCount to vertexRefCount, since it gets modified. I do not know whether or not I will need it or not.
Refactor incompletePolygons into an array of indices, and use indices for PolygonJoins and PolygonJoins.otherPolyJoins
Join polygons with connected start/end vertices first
Perform the joins by removing the polygons at the given indices from sectorPolygons, and inserting the new completed polygons, except if said polygons are incomplete.