Decouple 3D map builder from 3D view
The 3D map builder was designed to create a 3D version of a Doom level for viewing in 3D with Three.js. This decouples the 3D map builder from the view so that it can be more easily converted into other formats, such as OBJ. Unfortunately, this means the 3D view currently does not work.