Fix midtexture height and UV calculations again
Hopefully, midtexture UV and height calculations are fixed for good this time. I tested this with a map where tall midtextures were set up to look like a sine wave, going both above and below the ceiling and floor respectively.

I may move the quad heights recalculations to a static method in MapGeometryBuilder.