Attempt to make 3D map builder more fault tolerant
Attempt to improve the way edges and vertices are iterated over by checking whether or not all of the edges have been added.
Treat two-sided linedefs with a back sidedef of 0xffff as one-sided linedefs, since there are some maps that have that bug for some reason.
Return nothing if a start edge for the sector cannot be found.
Rework edge checks.
Remove useless code that adds edge vertices to sectorVertexIndices.
Gracefully handle cases where sector polygons cannot be built instead of not handling the error.