Attempt to fix 3D view of Eviternity MAP27 and MAP29
Add debug variable to SectorPolygonBuilder so that it knows when to print out extra information
Flip the Y coordinates of vertices, since that's the way to convert between Doom map coordinates and THREE.js coordinates.
In addition to picking the rightmost vertex for the edge start vertex, ensure the vertex is also the uppermost vertex if there are multiple vertices at the same X coordinate.
Remove useless else block from findNextVertex, since nothing else comes after it
Improve isPolygonComplete so that it checks to see whether or not the edge is used if it exists, and return true if it is not used and it exists
Add conditional debug logging to getPolygons()
Do not mark a polygon as a hole if there is a vertex that is at the same position as a vertex on another polygon