Fix Ancient Aliens MAP29 and Heretic E1M9 3D view
Filter the list of sector polygons so that incomplete polygons without an edge connecting the start and end vertex will not be in the list of complete polygons.
Ensure the edge between lastVertex and nextVertex exists, and use it. Otherwise, the polygon is a dud and should be removed.
Log the error itself if a sector polygon cannot be built.
Remove unused code to get a material asynchronously.