Add WebVR support and VR controls
Add a THREE.js "WebVR" button to allow VR headset users to view Doom maps in VR.
Add "viewHeadMoving" variable, which is turned on and off by events emitted by the controller.
Move the viewer forward when the controller button is pressed.
NOTE: This was only tested on a Samsung Gear VR headset.
Also, remove wireframe 3D view because it doesn't work right now, and I needed to have enough space on the VR browser window in order to activate the 3D map view.