Rethink next vertex and hole selection
The way I did next vertex selection always selected vertices on the interior rather than the exterior. For the ear clipping algorithm to work properly, it seems hole contours have to avoid overlapping. Now, interior or exterior vertices are selected based on whether they are inside another polygon or not, as determined by the point in polygon algorithm

Add logVertices method to SectorPolygonBuilder, which logs the previous, current, and next vertices
Move pointInPolygon outside of MapGeometryBuilder, since it is used nowhere else in the code
Change return type of SectorPolygonBuilder.findNextStartEdge to [number, number], since this is needed to "apply" this.visitEdge on nextStartEdge
Rename "clockwise" to "exterior" in some places
Add edgeInPolygon method, which checks whether an edge is inside a polygon or not
Rethink polygon hole selection to check whether any of the polygon's unique vertices are inside another polygon, and whether the number of container polygons is odd
Log information about the point in polygon stuff if debug mode is turned on
Add MAP12 to 3difytest.wad