Add support for different material styles
Add getColormaps method to TextureLibrary, which gets the palette, and the colormaps, as an RGBA format pixel buffer.
Add a buffer and attribute for the per-vertex sector light level, which is used by the software-style material.
Add MaterialStyle enum to three.ts, which specifies the desired appearance of the material.
Use the materialStyle value to determine which values to set for which variables and constants before creating the texture and material.
Add colormapTexture protected field to BufferModel.
Rename createIndexedShaderMaterial to createSoftwareStyleMaterial, and make it a private static method of BufferModel.