Add support for Doom 64 style rendering
Since the original Doom 64 ran on the Nintendo 64, Doom 64 uses the N64's three-point bilinear filtering. I got the shader code for it from ArthurCarvalho on the EmuTalk forums, ported it to GLSL, and modified it to be compatible with the 3D view shader.

I also included a function which mirrors the UV coordinates on the X and/or Y axis.

Also, change light level buffer to an Int32 array, since that seems to be the only way to get WebGL to recognize it as an int.